src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/remove-shadowroot-from-document-and-destroy-crash-expected.txt

CONSOLE ERROR: line 9: Uncaught TypeError: div1.createShadowRoot is not a function
